refactor: code decomposition
