+++ src/helpers/getRootDir.js
@@ -1,0 +1,6 @@
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+
+const getRootDir = (url) => dirname(fileURLToPath(url));
+
+export default getRootDir;
